Implemented initial schema parsing and graph construction.
Currently graph  creation from a single  data source (CSV file)  is supported.
The source(s) specified in the schema are  ignored and the actual data file is
provided to  the Ingestor  as command  line parameter.  Ingestion can  only be
started via the command line (i.e. no REST interface for the time being).

The ingestion process results in output  of the ingested graph to /tmp/<label>
in EPGM-JSON and GDF formats.

Included under  resources/examples is a  small imdb (movie  database) dataset,
containing a single CSV file and the corresponding schema.