Directory structure for documentation and missing features.
The documentation uses GitHub's markdown format. The README.md, which gives an
overview,  now   links  to  doc/index.md,   the  entry  point  for   the  full
documentation.

File   doc/index.md   contains   a   table  of   contents   to   documentation
topics (currently only stubs).

A file  TODO.md was added  to list and  prioritize missing features  that have
been identified, but not yet implemented.  The file is listed in .gitignore to
make its status independent from the current branch.